create EiffelFlowConsoleCLI
